Modularize greater.
