+++ tfjs-core/src/ops/ops.ts
@@ -32,6 +32,7 @@ export {concat4d} from './concat_4d';
 export {div} from './div';
 export {divNoNan} from './div_no_nan';
 export {eye} from './eye';
+export {greater} from './greater';
 export {multinomial} from './multinomial';
 export {notEqual} from './not_equal';
 export {oneHot} from './one_hot';
